303: move JsonSchemaEntities to akkahttp-server-circe build, add akkahttp-server-playjson
