+++ akka-http/build.sbt
@@ -5,10 +5,12 @@ val `algebra-circe-jvm` = LocalProject("algebra-circeJVM")
 val `algebra-playjson-jvm` = LocalProject("algebra-playjsonJVM")
 val `json-schema-circe-jvm` = LocalProject("json-schema-circeJVM")
 val `json-schema-generic-jvm` = LocalProject("json-schema-genericJVM")
+val `json-schema-playjson-jvm` = LocalProject("json-schema-playjsonJVM")
 
 val akkaActorVersion = "2.5.22"
 val akkaHttpVersion = "10.1.8"
 val akkaHttpCirceVersion = "1.25.2"
+val akkaHttpJsonVersion = "1.25.2"
 
 val `akka-http-client` =
   project.in(file("client"))
@@ -37,6 +39,20 @@ val `akka-http-server` =
       libraryDependencies ++= Seq(
         "com.typesafe.akka" %% "akka-stream" % akkaActorVersion,
         "com.typesafe.akka" %% "akka-http" % akkaHttpVersion,
+        scalaTestDependency
+      )
+    )
+    .dependsOn(`algebra-jvm` % "test->test;compile->compile")
+
+val `akka-http-server-circe` =
+  project.in(file("server-circe"))
+    .settings(
+      publishSettings,
+      `scala 2.11 to 2.12`,
+      name := "endpoints-akka-http-server-circe",
+      libraryDependencies ++= Seq(
+        "com.typesafe.akka" %% "akka-stream" % akkaActorVersion,
+        "com.typesafe.akka" %% "akka-http" % akkaHttpVersion,
         "com.typesafe.akka" %% "akka-http-testkit" % akkaHttpVersion % Test,
         "com.typesafe.akka" %% "akka-testkit" % akkaActorVersion % Test,
         "de.heikoseeberger" %% "akka-http-circe" % akkaHttpCirceVersion % Provided,
@@ -48,3 +64,26 @@ val `akka-http-server` =
     .dependsOn(`algebra-circe-jvm` % "test->test")
     .dependsOn(`json-schema-circe-jvm`)
     .dependsOn(`json-schema-generic-jvm` % "test->test")
+    .dependsOn(`akka-http-server`, `algebra-circe-jvm`, `json-schema-circe-jvm`)
+
+val `akka-http-server-playjson` =
+  project.in(file("server-playjson"))
+    .settings(
+      publishSettings,
+      `scala 2.11 to 2.12`,
+      name := "endpoints-akka-http-server-playjson",
+      libraryDependencies ++= Seq(
+        "com.typesafe.akka" %% "akka-stream" % akkaActorVersion,
+        "com.typesafe.akka" %% "akka-http" % akkaHttpVersion,
+        "com.typesafe.akka" %% "akka-http-testkit" % akkaHttpVersion % Test,
+        "com.typesafe.akka" %% "akka-testkit" % akkaActorVersion % Test,
+        "de.heikoseeberger" %% "akka-http-play-json" % akkaHttpJsonVersion,
+        "com.typesafe.play" %% "play-json" % playjsonVersion,
+        scalaTestDependency
+      )
+    )
+    .dependsOn(`algebra-jvm` % "test->test;compile->compile")
+    .dependsOn(`json-schema-playjson-jvm`)
+    .dependsOn(`json-schema-generic-jvm` % "test->test")
+    .dependsOn(`akka-http-server`)
+
