+++ akka-http/server-circe/src/main/scala/endpoints/akkahttp/server/circe/JsonSchemaEntities.scala
@@ -1,10 +1,11 @@
-package endpoints.akkahttp.server
+package endpoints.akkahttp.server.circe
 
 import akka.http.scaladsl.server.Directives
 import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
 import endpoints._
 import endpoints.akkahttp.server
 import endpoints.algebra.Documentation
+
 import io.circe.{Decoder, Encoder}
 
 /**
