+++ akka-http/server/src/main/scala/endpoints/akkahttp/server/JsonEntitiesFromCodec.scala
@@ -1,4 +1,5 @@
 package endpoints.akkahttp.server
+
 import akka.http.scaladsl.marshalling.{Marshaller, ToEntityMarshaller}
 import akka.http.scaladsl.model.{HttpEntity, MediaTypes}
 import akka.http.scaladsl.server.Directives
